worked on game class
started trying to replace the OldGame class (formerly Game) with Game (formerly NewGame) in all places where it is used. Lots of errors atm
