+++ Game.ctxt
@@ -1,181 +1,47 @@
 #BlueJ class context
 comment0.target=Game
-comment1.params=teamA\ teamB\ dayNum\ startTime\ season
-comment1.target=Game(Team,\ Team,\ int,\ long,\ int)
+comment1.params=teamA\ teamB\ dayNum\ startTime
+comment1.target=Game(Team,\ Team,\ int,\ long)
 comment10.params=
-comment10.target=void\ beforeOut()
+comment10.target=boolean\ strikeout()
 comment11.params=
-comment11.target=void\ afterBatterDeclared()
+comment11.target=java.lang.String\ bs()
 comment12.params=
-comment12.target=void\ afterPitch()
+comment12.target=void\ clearBases()
 comment13.params=
-comment13.target=void\ beforeHalfInning()
+comment13.target=void\ ball()
 comment14.params=
-comment14.target=void\ afterInning()
+comment14.target=void\ walk()
 comment15.params=
-comment15.target=void\ afterHalfInning()
-comment16.params=
-comment16.target=double\ getScoreA()
-comment17.params=
-comment17.target=double\ getScoreB()
+comment15.target=void\ printScore()
+comment16.params=score
+comment16.target=java.lang.String\ scoreToString(double)
+comment17.params=p
+comment17.target=void\ score(Player)
 comment18.params=
-comment18.target=Team\ getWinner()
-comment19.params=p
-comment19.target=void\ score(Player)
+comment18.target=boolean\ endOfGame()
+comment19.params=
+comment19.target=boolean\ endOfInning()
 comment2.params=
-comment2.target=void\ runGame()
-comment20.params=basesRun
-comment20.target=void\ advanceBaserunners(int)
-comment21.params=base\ basesRun
-comment21.target=boolean\ canAdvanceExtra(int,\ int)
-comment22.params=base
-comment22.target=java.lang.String\ getBaseName(int)
-comment23.params=
-comment23.target=boolean\ maximumBlaseball()
-comment24.params=
-comment24.target=boolean\ basesLoaded()
-comment25.params=
-comment25.target=void\ walk()
-comment26.params=
-comment26.target=void\ printScore()
-comment27.params=score
-comment27.target=java.lang.String\ scoreToString(double)
-comment28.params=input
-comment28.target=java.lang.String\ cloud(java.lang.String)
-comment29.params=message
-comment29.target=void\ printC(java.lang.String)
+comment2.target=void\ simulateGame()
+comment20.params=
+comment20.target=void\ setNextBatter()
+comment21.params=s\ tick
+comment21.target=void\ addEvent(java.lang.String,\ int)
+comment22.params=
+comment22.target=void\ setRandomWeather()
 comment3.params=
-comment3.target=void\ out()
-comment30.params=message
-comment30.target=void\ printlnC(java.lang.String)
-comment31.params=message
-comment31.target=void\ printInstantC(java.lang.String)
-comment32.params=message
-comment32.target=void\ printlnInstantC(java.lang.String)
-comment33.params=message
-comment33.target=void\ print(java.lang.String)
-comment34.params=message
-comment34.target=void\ println(java.lang.String)
-comment35.params=message
-comment35.target=void\ printInstant(java.lang.String)
-comment36.params=message
-comment36.target=void\ printlnInstant(java.lang.String)
-comment37.params=time
-comment37.target=void\ wait(int)
-comment38.params=
-comment38.target=java.lang.String\ gameName()
-comment39.params=
-comment39.target=boolean\ isLive()
-comment4.params=
-comment4.target=void\ beforeWins()
-comment40.params=
-comment40.target=boolean\ hasStarted()
-comment41.params=
-comment41.target=void\ watch()
-comment42.params=
-comment42.target=void\ leave()
-comment43.params=
-comment43.target=java.lang.String\ randomWeather()
-comment44.params=
-comment44.target=java.lang.String\ horribleJoke()
-comment45.params=
-comment45.target=java.lang.String\ randomTimelessWeather()
-comment46.params=
-comment46.target=java.lang.String\ randomAltWeather()
-comment47.params=w
-comment47.target=boolean\ weatherValid(java.lang.String)
-comment48.params=
-comment48.target=void\ printBirdMessage()
-comment49.params=
-comment49.target=void\ printFaradayMessage()
-comment5.params=p\ base
-comment5.target=void\ afterEnterBase(Player,\ int)
-comment50.params=
-comment50.target=void\ friendship()
-comment51.params=
-comment51.target=Team\ randomTeam()
-comment52.params=
-comment52.target=void\ printFeverMessage()
-comment53.params=base
-comment53.target=java.lang.String\ shortBaseName(int)
-comment54.params=p\ base
-comment54.target=void\ faradayStrike(Player,\ int)
-comment55.params=distanceRun
-comment55.target=void\ geometricWalk(double)
-comment56.params=distanceRun
-comment56.target=void\ rotateBases(double)
-comment57.params=
-comment57.target=void\ errorOut()
-comment58.params=
-comment58.target=void\ errorHalfInning()
-comment59.params=
-comment59.target=void\ errorEvent()
-comment6.params=
-comment6.target=void\ nextBatter()
-comment60.params=
-comment60.target=boolean\ scoreWillFlip()
-comment61.params=
-comment61.target=void\ shuffleBases()
-comment62.params=
-comment62.target=void\ spaceOut()
-comment63.params=p\ team
-comment63.target=Player\ nextAlphabetically(Player,\ Team)
-comment64.params=players
-comment64.target=Player[]\ sortAlphabetically(Player[])
-comment65.params=a\ b
-comment65.target=boolean\ AhigherThanB(Player,\ Player)
-comment66.params=
-comment66.target=void\ pixelize()
-comment67.params=num
-comment67.target=int\ roundToNearestOddNumber(double)
-comment68.params=
-comment68.target=void\ lightningStrike()
-comment69.params=
-comment69.target=void\ electricStrike()
-comment7.params=p
-comment7.target=void\ removePlayerFromBases(Player)
-comment70.params=
-comment70.target=void\ startDuel()
-comment71.params=
-comment71.target=void\ moonEvent()
-comment72.params=
-comment72.target=void\ popBubble()
-comment73.params=
-comment73.target=void\ normalStatChange()
-comment74.params=
-comment74.target=void\ preOutCatEvent()
-comment75.params=
-comment75.target=void\ catEvent()
-comment76.params=
-comment76.target=Player\ randomCatDefender()
-comment77.params=t
-comment77.target=int\ playersPresent(Team)
-comment78.params=
-comment78.target=boolean\ playersElsewhere()
-comment79.params=
-comment79.target=Player\ randomPresentDefender()
-comment8.params=
-comment8.target=void\ clearBases()
-comment80.params=
-comment80.target=Player\ randomCatOnBase()
-comment81.params=
-comment81.target=void\ printElectroMessage()
-comment82.params=p
-comment82.target=void\ electroBoost(Player)
-comment83.params=
-comment83.target=void\ feverPitch()
-comment84.params=
-comment84.target=int\ randomUnoccupiedBase()
-comment85.params=team
-comment85.target=Player\ randomActivePlayer(Team)
-comment86.params=
-comment86.target=Player\ randomActivePlayer()
-comment87.params=
-comment87.target=Player[]\ activePlayers()
-comment88.params=lineup\ p
-comment88.target=Player[]\ removePlayer(Player[],\ Player)
-comment89.params=p
-comment89.target=Player[]\ copy(Player[])
+comment3.target=void\ giveWins()
+comment4.params=baseNum
+comment4.target=void\ stealAttempt(int)
+comment5.params=
+comment5.target=void\ doSteals()
+comment6.params=p
+comment6.target=boolean\ wantsToSteal(Player)
+comment7.params=
+comment7.target=void\ doPitch()
+comment8.params=basesRun
+comment8.target=void\ advanceBaserunners(int)
 comment9.params=
-comment9.target=void\ halfInningStart()
-numComments=90
+comment9.target=Player\ randomDefender()
+numComments=23
