+++ League.ctxt
@@ -14,9 +14,9 @@ comment14.params=
 comment14.target=void\ scheduleSeason()
 comment15.params=startTime
 comment15.target=void\ scheduleSeason(long)
+comment15.text=public\ static\ void\ scheduleSeason(long\ startTime){//timeless\ league\ lol\r\n\ \ \ \ \ \ \ \ seasonGames\ \=\ new\ Game[1][1];\r\n\ \ \ \ \ \ \ \ seasonGames[0][0]\ \=\ new\ Game(ultraDark[0],ultraDark[1],1,seasonStartTime,season);\ \r\n}/*
 comment16.params=
 comment16.target=void\ schedulePostseason1()
-comment16.text=public\ static\ void\ scheduleSeason(long\ startTime){\r\n\ \ \ \ \ \ \ \ seasonGames\ \=\ new\ Game[99][getTeams().length\ /\ 2];\r\n\ \ \ \ \ \ \ \ for(int\ x\ \=\ 0;\ x\ <\ 99;\ x+\=3){\r\n\ \ \ \ \ \ \ \ \ \ \ \ ArrayList<Team>\ darkTeams\ \=\ new\ ArrayList<Team>();\r\n\ \ \ \ \ \ \ \ \ \ \ \ ArrayList<Team>\ lightTeams\ \=\ new\ ArrayList<Team>();\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(int\ y\ \=\ 0;\ y\ <\ ultraDark.length;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ darkTeams.add(ultraDark[y]);\r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(int\ y\ \=\ 0;\ y\ <\ moderateDark.length;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ darkTeams.add(moderateDark[y]);\r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(int\ y\ \=\ 0;\ y\ <\ ultraLight.length;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ lightTeams.add(ultraLight[y]);\r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(int\ y\ \=\ 0;\ y\ <\ moderateLight.length;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ lightTeams.add(moderateLight[y]);\r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ \ \ \ \ int\ y\ \=\ 0;\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(;darkTeams.size()\ >\ 1;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ int\ r1\ \=\ (int)(r.nextDouble()\ *\ darkTeams.size());\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ Team\ t1\ \=\ darkTeams.get(r1);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ darkTeams.remove(r1);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ int\ r2\ \=\ (int)(r.nextDouble()\ *\ darkTeams.size());\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ Team\ t2\ \=\ darkTeams.get(r2);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ darkTeams.remove(r2);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x][y]\ \=\ new\ Game(t1,t2,x+1,startTime\ +\ (3600000\ *\ x),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x+1][y]\ \=\ new\ Game(t1,t2,x+2,startTime\ +\ (3600000\ *\ (x+1)),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x+2][y]\ \=\ new\ Game(t1,t2,x+3,startTime\ +\ (3600000\ *\ (x+2)),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ \ \ \ \ for(;lightTeams.size()\ >\ 1;\ y++){\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ int\ r1\ \=\ (int)(r.nextDouble()\ *\ lightTeams.size());\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ Team\ t1\ \=\ lightTeams.get(r1);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ lightTeams.remove(r1);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ int\ r2\ \=\ (int)(r.nextDouble()\ *\ lightTeams.size());\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ Team\ t2\ \=\ lightTeams.get(r2);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ lightTeams.remove(r2);\r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x][y]\ \=\ new\ Game(t1,t2,x+1,startTime\ +\ (3600000\ *\ x),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x+1][y]\ \=\ new\ Game(t1,t2,x+2,startTime\ +\ (3600000\ *\ (x+1)),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ \ \ \ \ seasonGames[x+2][y]\ \=\ new\ Game(t1,t2,x+3,startTime\ +\ (3600000\ *\ (x+2)),season);\ \r\n\ \ \ \ \ \ \ \ \ \ \ \ }\r\n\ \ \ \ \ \ \ \ }\r\n}
 comment17.params=
 comment17.target=void\ schedulePostseason2()
 comment18.params=
@@ -43,22 +43,20 @@ comment27.params=
 comment27.target=void\ runAllPostseason3Games()
 comment28.params=season
 comment28.target=Game[][]\ getGames(int)
-comment29.params=games\ season
-comment29.target=void\ setSeasonOfGames(Game[][],\ int)
+comment29.params=
+comment29.target=void\ recapSeason0()
 comment3.params=day
 comment3.target=void\ runPostseason(int)
 comment30.params=
-comment30.target=void\ recapSeason0()
+comment30.target=void\ recapSeason3()
 comment31.params=
-comment31.target=void\ recapSeason3()
+comment31.target=void\ recapSeason1()
 comment32.params=
-comment32.target=void\ recapSeason1()
+comment32.target=void\ recapSeason2()
 comment33.params=
-comment33.target=void\ recapSeason2()
+comment33.target=void\ recapVortex()
 comment34.params=
-comment34.target=void\ recapVortex()
-comment35.params=
-comment35.target=void\ recapAltAlt()
+comment34.target=void\ recapAltAlt()
 comment4.params=day
 comment4.target=void\ runGamesUpToDay(int)
 comment5.params=
@@ -71,4 +69,4 @@ comment8.params=
 comment8.target=int\ longestTeamNameLength()
 comment9.params=s\ length
 comment9.target=java.lang.String\ addSpaces(java.lang.String,\ int)
-numComments=36
+numComments=35
