--- NewGame.ctxt
@@ -1,37 +1,0 @@
-#BlueJ class context
-comment0.target=NewGame
-comment1.params=teamA\ teamB\ dayNum\ startTime
-comment1.target=NewGame(Team,\ Team,\ int,\ long)
-comment10.params=
-comment10.target=void\ printScore()
-comment11.params=score
-comment11.target=java.lang.String\ scoreToString(double)
-comment12.params=p
-comment12.target=void\ score(Player)
-comment13.params=
-comment13.target=boolean\ endOfGame()
-comment14.params=
-comment14.target=boolean\ endOfInning()
-comment15.params=
-comment15.target=void\ setNextBatter()
-comment16.params=s\ tick
-comment16.target=void\ addEvent(java.lang.String,\ int)
-comment17.params=
-comment17.target=void\ setRandomWeather()
-comment2.params=
-comment2.target=void\ simulateGame()
-comment3.params=basesRun
-comment3.target=void\ advanceBaserunners(int)
-comment4.params=
-comment4.target=Player\ randomDefender()
-comment5.params=
-comment5.target=boolean\ strikeout()
-comment6.params=
-comment6.target=java.lang.String\ bs()
-comment7.params=
-comment7.target=void\ clearBases()
-comment8.params=
-comment8.target=void\ ball()
-comment9.params=
-comment9.target=void\ walk()
-numComments=18
